TIP-835: Add js acceptance tests to jenkinsfile
TIP-835: Ignore ACl in form extensions dump

Create test_dist folder in jenkinsfile

Disable sandbox args

Enable sandbox args

Run acceptance test in npm

Build

Add reporter script

Change image in withNode

Fix race condition in page/browser close

Remove step-dictionary and use glob

Refactor run.js

TIP-835: Use custom node docker image
