+++ package.json
@@ -10,8 +10,9 @@
     "webpack-test": "webpack --config webpack-test.config.js",
     "lint": "node ./node_modules/eslint/bin/eslint 'src/**/*.js' --ignore-path .eslintignore --quiet",
     "lint-fix": "node ./node_modules/eslint/bin/eslint 'src/**/*.js' --ignore-path .eslintignore --quiet --fix",
-    "acceptance": "cucumber-js -r ./webpack/test/acceptance/run.js -r ./tests/front/acceptance/cucumber ./tests/front/acceptance/features/",
-    "acceptance-debug": "cucumber-js -r ./webpack/test/acceptance/run.js -r ./tests/front/acceptance/cucumber/step-definitions ./tests/front/acceptance/features/ --world-parameters='{\"debug\": true}'"
+    "acceptance": "cucumber-js -r ./webpack/test/acceptance/run-steps.js -r ./tests/front/acceptance/cucumber ./tests/front/acceptance/features/",
+    "acceptance-debug": "cucumber-js -r ./webpack/test/acceptance/run-steps.js -r ./tests/front/acceptance/cucumber/step-definitions ./tests/front/acceptance/features/ --world-parameters='{\"debug\": true}'",
+    "acceptance-report": "yarn run acceptance -- -f json:web/test_dist/acceptance-js.json && node webpack/test/acceptance/generate-report.js"
   },
   "repository": {
     "type": "git",
@@ -26,6 +27,7 @@
     "cucumber": "^4.0.0",
     "cucumber-html-reporter": "^4.0.1",
     "dep-diff": "1.0.1",
+    "glob": "^7.1.2",
     "html-webpack-inline-source-plugin": "^0.0.9",
     "html-webpack-plugin": "^2.30.1",
     "puppeteer": "^1.1.1",
@@ -55,7 +57,6 @@
     "merge-objects": "1.0.5",
     "raw-loader": "0.5.1",
     "regexp-replace-loader": "1.0.0",
-    "step-dictionary": "^0.1.2",
     "summernote": "0.6.5",
     "text-loader": "0.0.1",
     "underscore": "1.8.3",
